# Get rid of some viper.Get* calls