@bep 2017-02-20
Pretty good effect for small amount of work in my benchmark:

```
benchmark           old ns/op      new ns/op      delta
BenchmarkHugo-4     9623753236     9264761669     -3.73%

benchmark           old allocs     new allocs     delta
BenchmarkHugo-4     43342841       42696420       -1.49%

benchmark           old bytes      new bytes      delta
BenchmarkHugo-4     9084994152     9065991096     -0.21%
```