@bogem 2017-02-20
Hm, it's very interesting, how viper really badly affects on the performance and memory consumption.