@bep 2017-02-20
Yes, but according to your latest numbers, we have removed the bottle necks now.

This is the benchmarks from the Viper project:

```
BenchmarkGetBool-4          	 3000000	       426 ns/op	      65 B/op	       5 allocs/op
BenchmarkGet-4              	 5000000	       373 ns/op	      64 B/op	       5 allocs/op
BenchmarkGetBoolFromMap-4   	200000000	         6.36 ns/op	       0 B/op	       0 allocs/op
````
And while 400 ns is 60 times slower than a map lookup, you would need lots of them to notice.

Which means, we should continue to clean up -- but only if it actually gets cleaner, not by making lots of local variables around the place.