@bogem 2017-02-20
Ok, I understood you. I will try to just enforce PathSpecs or delete really redundant viper calls.